+++ src/main/java/com/wwinfo/service/impl/RfidreaderServiceImpl.java
@@ -1,11 +1,14 @@
 package com.wwinfo.service.impl;
 
 import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.core.collection.CollUtil;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.wwinfo.common.ExcludeEmptyQueryWrapper;
 import com.wwinfo.common.exception.BusinessException;
+import com.wwinfo.mapper.EntryposMapper;
+import com.wwinfo.model.Entrypos;
 import com.wwinfo.model.Rfidreader;
 import com.wwinfo.mapper.RfidreaderMapper;
 import com.wwinfo.pojo.query.RfidreaderQuery;
@@ -17,6 +20,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
+import java.util.List;
 
 /**
  * <p>
@@ -31,6 +35,9 @@ public class RfidreaderServiceImpl extends ServiceImpl<RfidreaderMapper, Rfidrea
 
     @Resource
     private RfidreaderMapper rfidreaderMapper;
+
+    @Resource
+    private EntryposMapper entryposMapper;
 
     @Override
     public IPage listPage(RfidreaderQuery rfidreaderQuery) {
@@ -65,8 +72,24 @@ public class RfidreaderServiceImpl extends ServiceImpl<RfidreaderMapper, Rfidrea
     @Override
     public int delete(Long id) {
         if(id == null)
-            throw new BusinessException("id不能重复");
+            throw new BusinessException("id不能为空");
+        //校验阅读器是否在出入点配置中使用
+        QueryWrapper<Entrypos> wrapper = new ExcludeEmptyQueryWrapper<>();
+        wrapper.eq("startReaderID", id)
+                .or()
+                .eq("endReaderID", id);
+        List<Entrypos> entryposList = entryposMapper.selectList(wrapper);
+        if(CollUtil.isNotEmpty(entryposList)){
+            throw new BusinessException("该阅读器已在出入点配置中使用,不能删除");
+        }
         return rfidreaderMapper.deleteById(id);
+    }
+
+    @Override
+    public Rfidreader getByReaderName(String readerName) {
+        QueryWrapper<Rfidreader> wrapper = new ExcludeEmptyQueryWrapper<>();
+        wrapper.eq("readerName", readerName);
+        return rfidreaderMapper.selectOne(wrapper);
     }
